refactor(memory): consolidate memory operations into modular structure with new document, file, and key-value handlers
